move theme to main dir
