+++ vue-rewrite/src/theme.ts
@@ -1,54 +1,109 @@
-export const defaultSizeVariables = {
-  "icon-size": "2em",
-  "border-radius": "0.5rem",
-}
+import { definePreset } from "@primevue/themes"
+import Aura from "@primevue/themes/aura"
 
-export function getTheme(theme: "dark" | "light", options = { variables: { ...defaultSizeVariables } }) {
-  if (theme === "light") {
-    return {
-      dark: false,
-      colors: {
-        primary: "#212121",
-        secondary: "#212121",
-        accent: "#212121",
-        error: "#d32e2e",
-        onError: "#ffffff",
-        warning: "#f47a00",
-        onWarning: "#000000",
-        info: "#0288d1",
-        onInfo: "#000000",
-        success: "#378e3c",
-        onSuccess: "#ffffff",
+export const themePreset = definePreset(Aura, {
+  components: {
+    button: {
+      paddingX: "0.75rem",
+      paddingY: "0.5rem",
+    },
+    select: {
+      colorScheme: {
+        light: {
+          optionColor: "{surface.700}",
+        },
+        dark: {
+          optionColor: "{surface.300}",
+        },
       },
-      variables: options.variables,
-    }
-  }
-  return {
-    dark: true,
-    colors: {
-      primary: "#212121",
-      secondary: "#212121",
-      accent: "#212121",
-      error: "#d32e2e",
-      onError: "#ffffff",
-      warning: "#f47a00",
-      onWarning: "#000000",
-      info: "#0288d1",
-      onInfo: "#000000",
-      success: "#378e3c",
-      onSuccess: "#ffffff",
     },
-    variables: {
-      "overlay-opacity": 0.15,
-      "activated-opacity": 0.4,
-      "theme-overlay-multiplier": 1,
-      "theme-active-background": "var(--v-primary-base)",
-      "theme-active-color": "var(--v-primary-base)",
-      ...options.variables,
+  },
+  semantic: {
+    primary: {
+      50: "{zinc.50}",
+      100: "{zinc.100}",
+      200: "{zinc.200}",
+      300: "{zinc.300}",
+      400: "{zinc.400}",
+      500: "{zinc.500}",
+      600: "{zinc.600}",
+      700: "{zinc.700}",
+      800: "{zinc.800}",
+      900: "{zinc.900}",
+      950: "{zinc.950}",
     },
-  }
-}
-
-export const lightTheme = getTheme("light")
-
-export const darkTheme = getTheme("dark")
+    colorScheme: {
+      light: {
+        primary: {
+          color: "{zinc.900}",
+          inverseColor: "#ffffff",
+          hoverColor: "{zinc.900}",
+          activeColor: "{zinc.800}",
+        },
+        surface: {
+          50: "{gray.50}",
+          100: "{gray.100}",
+          200: "{gray.200}",
+          300: "{gray.300}",
+          400: "{gray.400}",
+          500: "{gray.500}",
+          600: "{gray.600}",
+          700: "{gray.700}",
+          800: "{gray.800}",
+          900: "{gray.900}",
+          950: "{gray.950}",
+        },
+        text: {
+          color: "{surface.700}",
+          hoverColor: "{surface.800}",
+          mutedColor: "{surface.500}",
+          hoverMutedColor: "{surface.600}",
+        },
+        highlight: {
+          background: "{zinc.950}",
+          focusBackground: "{zinc.700}",
+          color: "#ffffff",
+          focusColor: "#ffffff",
+        },
+      },
+      dark: {
+        primary: {
+          color: "{zinc.50}",
+          inverseColor: "{zinc.950}",
+          hoverColor: "{zinc.100}",
+          activeColor: "{zinc.200}",
+        },
+        surface: {
+          50: "{gray.50}",
+          100: "{gray.100}",
+          200: "{gray.200}",
+          300: "{gray.300}",
+          400: "{gray.400}",
+          500: "{gray.500}",
+          600: "{gray.600}",
+          700: "{gray.700}",
+          800: "{gray.800}",
+          900: "{gray.900}",
+          950: "{gray.950}",
+        },
+        text: {
+          color: "{surface.700}",
+          hoverColor: "{surface.800}",
+          mutedColor: "{surface.500}",
+          hoverMutedColor: "{surface.600}",
+        },
+        highlight: {
+          background: "rgba(250, 250, 250, .16)",
+          focusBackground: "rgba(250, 250, 250, .24)",
+          color: "rgba(255,255,255,.87)",
+          focusColor: "rgba(255,255,255,.87)",
+        },
+        overlay: {
+          modal: {
+            color: "{surface.50}",
+          },
+        },
+      },
+    },
+  },
+})
